가게에 미션 조회 API 구현

+++ src/main/java/umc/study/web/dto/ReviewResponseDTO.java
@@ -14,7 +14,7 @@ public class ReviewResponseDTO {
         private Long storeId;
         private Long memberId;
         private String body;
-        private Double score;
+        private Float score;
         private LocalDateTime createdAt;
     }
 }
